Undid renaming/refactoring (as well as that Bind bugfix) to keep change footprint as small as possible
