+++ InedoCore/InedoExtension/UserDirectories/LdapClientSearchScope.cs
@@ -1,0 +1,9 @@
+﻿namespace Inedo.Extensions.UserDirectories
+{
+    internal enum LdapClientSearchScope
+    {
+        Base,
+        OneLevel,
+        Subtree
+    }
+}
